requirements.txt
# Terminal UI Framework
textual>=0.50.0
rich>=13.7.0

# Configuration
python-dotenv>=1.0.0

# Progress & Animation
yaspin>=3.0.0

# ASCII Art
pyfiglet>=1.0.0

# System monitoring
psutil>=5.9.0
